+++ tests/testthat/test-plot.check_outliers.R
@@ -13,4 +13,67 @@ test_that("`plot.see_check_outliers()` works", {
     plot(performance::check_outliers(model, verbose = FALSE)),
     "gg"
   )
+
+  skip_if_not_installed("ggrepel")
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works default method",
+    fig = plot(performance::check_outliers(model, verbose = FALSE))
+  )
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works bars method",
+    fig = suppressWarnings(plot(
+      performance::check_outliers(model, verbose = FALSE),
+      type = "bars"
+    ))
+  )
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works bars, rescaled",
+    fig = suppressWarnings(plot(
+      performance::check_outliers(model, verbose = FALSE),
+      type = "bars",
+      rescale_distance = TRUE
+    ))
+  )
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works count method",
+    fig = suppressWarnings(plot(
+      performance::check_outliers(model, verbose = FALSE),
+      type = "count"
+    ))
+  )
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works scree method",
+    fig = plot(performance::check_outliers(model, verbose = FALSE), type = "scree")
+  )
+
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers works scree method, z-score",
+    fig = plot(performance::check_outliers(mt2$mpg, method = "zscore"), type = "scree")
+  )
+})
+
+
+test_that("`plot.see_check_outliers()` multimethods", {
+  data(mtcars)
+  outliers_list <- performance::check_outliers(mtcars, method = c(
+    "mahalanobis",
+    "iqr",
+    "zscore"
+  ))
+  set.seed(123)
+  vdiffr::expect_doppelganger(
+    title = "plot.see_check_outliers multimethods",
+    fig = plot(outliers_list)
+  )
 })