+++ tests/testthat/test-plot.simulate_parameters.R
@@ -1,6 +1,4 @@
-skip_if(TRUE)
-
-## FIXME: currently does not retrieve the data
+skip("Fails at retrieving data")
 
 skip_on_cran()
 skip_if_offline()
@@ -10,11 +8,12 @@ skip_if_not_installed("parameters")
 
 test_that("`plot()` for simulate_parameters", {
   pict <- base::readRDS(url("https://slcladal.github.io/data/pict.rda", "rb"))
+  d <<- pict
   suppressWarnings({
     m1.mn <- mclogit::mblogit(
       formula = Response ~ Gender + Group,
       random = ~ 1 | Item,
-      data = pict
+      data = d
     )
   })
   set.seed(1234)